ci: Use regular Debian as base Windows for ARM
It turns out it's relatively easy to provide support for Windows on ARM.
Unfortunately, it requires building Wine from source, as the version
distributed by Debian (both Bookworm and Sid) is buggy for some tests.
However, building it isn't overly complicated, and can be replaced with
pre-built package once a fixed version lands in Debian Trixie.

Besides that, I also fixed the dynamic linking issue and enabled OpenMP
as for x86 targets.

It has the following advantages:

- It greatly decreases the image size, and improves maintainability of
  the package, as we don't rely on third-party image (from Linaro).
- We can build a native image and run it on ARM GitLab runner spreading
  the load between different runners, and significantly improving the
  speed of execution (over 12 min down to under a minute).

Signed-off-by: Marek Pikuła <m.pikula@partner.samsung.com>