RISC-V: fix rvv auto-detection on `gcc-13`
Without the change auto-detection of `rvv` support is incomplete as it
assumes `vfloat32m1x4_t` is always present if `vfloat32m1_t` is probed
successfully.

It looks like at least on `gcc-13.3.0` `vfloat32m1_t` is present while
`vfloat32m1x4_t` is not. This causes build failure as:

    ../pixman/pixman-rvv.c:464:5: error: unknown type name 'vfloat32m1x4_t'; did you mean 'vfloat32m4_t'?
      464 |     vfloat32m1x4_t sa_sr_sg_sb, da_dr_dg_db, ma_mr_mg_mb;
          |     ^~~~~~~~~~~~~~
          |     vfloat32m4_t

The change probes for both types as both are used in `pixman-rvv.c`.

Tested on `gcc-13`, `gcc-14`, `gcc-15`.

Closes: https://gitlab.freedesktop.org/pixman/pixman/-/issues/115