ci: Fix runner tag variable expansion
It seems that GitLab doesn't allow expansion of variables provided with
template "inputs" if they are not defined as CI variable on project
level. This makes the default values saves in global `variable` section
unusable. That means that if a projects doesn't have the CI variables
defined, the CI won't run – this includes forks.

In order to prevent problematic setup for forks, the default runner tags
are hard-coded.

Signed-off-by: Marek Pikuła <m.pikula@partner.samsung.com>