Fix arm64 advanced prefetcher
https://gitlab.freedesktop.org/pixman/pixman/-/issues/119

When the advanced preloader reached the end of a scanline, it advanced
the `PF_SRC`, `PF_DST`, and `PF_MASK` addresses by 1 instead of
advancing to the next scanline.

Most of the time, this means the prefetcher is simply prefetching the
wrong place in the image.  But when the stride is negative, this can
result in a segfault when trying to read memory beyond the end of the
image buffer.

I have fixed this by combining the bitshift and add instructions to
directly update the address register by the correct amount before
prefetching.

My fix results in the same behavior as the code in
`pixman/pixman-arm-neon-asm.h`:
```
    PF ldrbge, DUMMY, [PF_SRC, SRC_STRIDE, lsl #src_bpp_shift]!
```
This instruction means that `PF_SRC` is updated with the offset of
`SRC_STRIDE << #src_bpp_shift` and then the byte at updated `PF_SRC` is
read into `DUMMY`.  (The `ge` condition has been replaced with a
separate branch instruction in aarch64.)

I also cleaned up a couple other cases where instructions were
redundant.