Test case for compositing with a negative stride
The aarch64 advanced prefetcher has a bug where it can read past the
end of the image buffers.  Typically this causes no ill effects (beyond
making the attempted prefetch useless), because it's simply prefetching
junk data and discarding it.  Where this causes problems (i.e. segfault)
is when it tries to read memory that not readable, such as a memory map
with no read permission.

To expose this specific bug, we need a test case with a negative stride
and enough height such that `last_scanline_ptr + height` runs past the
end of the image buffer (this equation is approximate and omits some
small variables).  Using `fence_malloc`, we can catch the problem with a
segfault when the prefetch attempts to read memory beyond the end of the
image buffer.